feat: integrated GLTF loading
Integrated GLTF loading to the rest of pipeline. For draw calls, we have sceneMesh generated from the GLTF load as global Vertex and index buffers that contains all geometry data for the scene. The DrawTask descriptor now contains a scene::Primitive that defines offsets and counts inside these global buffers.